fix(web): make suggested tool JSON prettier-compliant and tested
- format clipboard JSON with prettier parser=json

- extract and reuse JSON builder utility

- add regression tests for newline and formatting compliance

- closes #445
