+++ web/package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
+    "test": "node --test src/utils/suggestToolJson.test.js",
     "lint": "eslint .",
     "format-json:check": "prettier --check ../quality-tools",
     "format-json:fix": "prettier --write ../quality-tools",
